Fix temporary table syntax for postgres.
Adding the fix for mysql/mariadb to use a temporary table
broke postgres.

Postgres does not support creating a temp
table using a select statement, unless it is done
through a select into, which maridb does not support.

This fix just wraps the original code (for postgres)
in a postgres engine check, and leaves the mariadb/mysql
code in that env.

Closes-Bug: 1800466